add `remove_on_delete` key
if :deletable and not :remove_on_delete, then deleting the row will set the value to nil. if :deletable AND :remove_on_delete, it actually removes the row from the UI

+++ lib/formotion/row_type/base.rb
@@ -49,9 +49,14 @@ module Formotion
 
       # called when the delete editing style was triggered tableView:commitEditingStyle:forRowAtIndexPath:
       def on_delete(tableView, tableViewDelegate)
-        row.section.rows.delete_at(row.index)
-        row.section.refresh_row_indexes
-        delete_row
+        if row.remove_on_delete?
+          row.section.rows.delete_at(row.index)
+          row.section.refresh_row_indexes
+          delete_row
+        else
+          row.value = nil
+          tableView.reloadData
+        end
       end
 
       def delete_row
